Renamed resume index.js to Resume.js
