+++ src/components/Resume/Resume.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Container from '@material-ui/core';
+import Container from '@material-ui/core/Container';
 
 import ResumePDF from './Roberto_Basulto_Resume.pdf'
